% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractpoly.R
\name{extractpoly}
\alias{extractpoly}
\title{Function to extract polygons}
\usage{
extractpoly(geosm_obj, place)
}
\arguments{
\item{geosm_obj}{An XML object downloaded from OpenStreetMap}

\item{place}{The place for which the XML information was downloaded}
}
\description{
Extract sp objects from XML data
}

