% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_osm.R
\name{geocode_osm}
\alias{geocode_osm}
\title{Function geocode_osm}
\usage{
geocode_osm(place, output = c("coord", "all"))
}
\arguments{
\item{place}{Place which should be geocoded}

\item{output}{Complete result or just the geocode}
}
\description{
Returns info from \url{http://nominatim.openstreetmap.org/}
}
\examples{
geocode_osm(place="Mannheim B2,1")
geocode_osm(place="Berlin",output="all")
}
\author{
Jan-Philipp Kolb <Jan-Philipp_Kolb@gmx.net>
}
\references{
\url{https://github.com/malexan/nominatim/blob/master/geocode.r}
}

