% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgeometry.R
\name{getgeometry}
\alias{getgeometry}
\title{Extract geometries from Openstreetmap}
\usage{
getgeometry(place)
}
\arguments{
\item{place}{Entity for which the geometry should be downloaded}
}
\value{
XML document with polygon from OpenStreetMap for the given place
}
\description{
Extract geometries from Openstreetmap using the Nominatim API
}
\examples{
MAseck <- getgeometry('Mannheim Seckenheim')
BE_geom <- getgeometry('Berlin')

}
\author{
Jan-Philipp Kolb <Jan-Philipp_Kolb@gmx.net>
}
\references{
\url{http://wiki.openstreetmap.org/wiki/Nominatim}
}
\seealso{
\code{\link{extractpoly}}
}

