\name{osm_df}
\alias{osm_df}
\title{data frame map features OpenStreetMap}
\usage{
data(osm_df)
}
\description{
Overview of all available map features at OpenStreetMap \url{http://wiki.openstreetmap.org/wiki/Map_Features}
}
\author{
Jan-Philipp Kolb \email{Jan-Philipp_Kolb@gmx.net}
}
\examples{
data(osm_df)
}
