% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_osm_nodes.R
\name{extract_osm_nodes}
\alias{extract_osm_nodes}
\title{function extract_osm_nodes}
\usage{
extract_osm_nodes(OSM.Data, value)
}
\arguments{
\item{OSM.Data}{an osm object}

\item{value}{a map feature}
}
\description{
function extract_osm_nodes
}
\examples{
schools <- get_osm_nodes(object='amenity=school',place='Heidelberg')
info <- extract_osm_nodes(OSM.Data=schools,value='school')
}

