% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_nodes.R
\name{get_osm_nodes}
\alias{get_osm_nodes}
\title{Function get_osm_nodes}
\usage{
get_osm_nodes(object, place)
}
\arguments{
\item{object}{A key-value combination for a node, valid combinations can be seen here: \url{http://wiki.openstreetmap.org/wiki/Map_Features}}

\item{place}{Place for the map features}
}
\value{
XML document with information for the nodes downloaded from OpenStreetMap for the given place
}
\description{
Extract objects from Openstreetmap using the Overpass API
}
\examples{
pg_MA <- get_osm_nodes(object="leisure=playground","Mannheim")
}
\author{
Jan-Philipp Kolb <Jan-Philipp_Kolb@gmx.net>
}
\references{
\url{http://wiki.openstreetmap.org/wiki/Map_Features}
}
\seealso{
\code{\link{extract_osm_nodes}}
}

