% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes4boundaries.R
\name{nodes4boundaries}
\alias{nodes4boundaries}
\title{function nodes4boundaries}
\usage{
nodes4boundaries(object, boundary)
}
\arguments{
\item{object}{A map feature}

\item{boundary}{A series of coordinates}
}
\value{
XML document with OpenStreetMap objects within the given boundary
}
\description{
Get OSM nodes within a given boundary
}
\examples{
objectI <- "leisure=playground"
boundaryI <- "50.6,7.0,50.8,7.3"
osm_obj <- nodes4boundaries(objectI,boundaryI)
}
\author{
Jan-Philipp Kolb <Jan-Philipp_Kolb@gmx.net>
}
\references{
\url{http://wiki.openstreetmap.org/wiki/Overpass_API}
}
\seealso{
\code{\link{extract_osm_nodes}}
}

